% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-coagulated.R
\name{bulk_peak_detect}
\alias{bulk_peak_detect}
\title{Get target peaks per group}
\usage{
bulk_peak_detect(df, diameter, full_df = FALSE)
}
\arguments{
\item{df}{`dataframe` of coerced measurements-`modules` as produced by `coulteR::bulk_read`.}

\item{diameter}{The expected diameter of particles}

\item{full_df}{Should all peaks be reported? default = `FALSE`}
}
\value{
A `dataframe`.
}
\description{
A wrapper around `coulteR::detect_peaks`. Per group in a dataframe with identifiers, peaks are reported.
}
