% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coulter-reader.R
\name{get_measurements}
\alias{get_measurements}
\title{Get Measurements from AccuComp-table}
\usage{
get_measurements(x)
}
\arguments{
\item{x}{filepath to AccuComp-table}
}
\description{
Load Measurements from AccuComp-table (last element in file).
importFrom stringr str_split
}
