% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucomp-reader.R
\name{get_measurements2}
\alias{get_measurements2}
\title{Get Measurements from AccuComp-table}
\usage{
get_measurements2(x)
}
\arguments{
\item{x}{filepath to AccuComp-table}
}
\description{
Load Measurements from AccuComp-table (last element in file).
}
