% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucomp-reader.R
\name{get_module2}
\alias{get_module2}
\title{Extract Summary Tables}
\usage{
get_module2(
  x,
  start,
  end,
  varnames,
  clean = FALSE,
  tidy = TRUE,
  trim = NULL,
  values_numeric = FALSE
)
}
\arguments{
\item{x}{Filepath to AccuComp-table (.XLS)}

\item{start}{Start-term to look for. Character vector one or two elements.}

\item{end}{end-term to look for. Character vector one or two elements.}

\item{varnames}{Desired column names of output.}

\item{clean}{Should output be cleaned from trailing ":"?}

\item{tidy}{Is input tidy (i.e. two variable names are present?)}

\item{trim}{Should output table be trimmed? Default NULL, else two-element numeric vector (specifying head and tail cutoff)}

\item{values_numeric}{Should values (second column) be transformed to numeric values?}
}
\description{
Extracts Summary tables from AccuComp-table
}
