% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-coagulated.R
\name{get_nearest_peak}
\alias{get_nearest_peak}
\title{Obtain peak nearest to target}
\usage{
get_nearest_peak(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{The target number}
}
\value{
A `number`
}
\description{
Obtain peak nearest to target
}
