% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-peaks.R
\name{get_number_cells}
\alias{get_number_cells}
\title{Get number of cells in range}
\usage{
get_number_cells(df, starts, ends)
}
\arguments{
\item{df}{A `measurements`-module}

\item{starts}{Indices of start bins}

\item{ends}{Indices of end bins}
}
\value{
A `vector` of summed cells
}
\description{
Get number of cells in range
}
