% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coulter-reader.R
\name{get_subtable}
\alias{get_subtable}
\title{Extract Summary Tables}
\usage{
get_subtable(x, start, end, varnames, clean = FALSE, tidy = TRUE)
}
\arguments{
\item{x}{Filepath to AccuComp-table (.XLS)}

\item{start}{Start-term to look for. Character vector one or two elements.}

\item{end}{end-term to look for. Character vector one or two elements.}

\item{varnames}{Desired column names of output.}

\item{clean}{Should output be cleaned from trailing ":"?}

\item{tidy}{Is input tidy (i.e. two variable names are present?)}
}
\description{
Extracts Summary tables from AccuComp-table
}
