% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-peaks.R
\name{get_wtd_mean}
\alias{get_wtd_mean}
\title{Get weighted means
Computes weighted means for all detected peaks}
\usage{
get_wtd_mean(df, peaks)
}
\arguments{
\item{df}{A `measurements`-module}

\item{peaks}{A `peaks` dataframe}
}
\value{
A numeric `vector`
}
\description{
Get weighted means
Computes weighted means for all detected peaks
}
