% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-peaks.R
\name{get_wtd_sd}
\alias{get_wtd_sd}
\title{Get weighted SDs
Computes weighted standard deviations for all detected peaks}
\usage{
get_wtd_sd(df, peaks)
}
\arguments{
\item{df}{A `measurements`-module}

\item{peaks}{A `peaks` dataframe}
}
\value{
A numeric `vector`
}
\description{
Get weighted SDs
Computes weighted standard deviations for all detected peaks
}
