% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-coagulated.R
\name{ggtracks}
\alias{ggtracks}
\title{Plot target tracks
Plots tracks of selected samples. Note that file extensions and the character `#` are to be removed from sample names!
This function either plots selected samples, or selected indices, or selects and plots 9 randomly selected tracks.}
\usage{
ggtracks(
  df,
  peaks = NULL,
  samples = NULL,
  seed = 123,
  N = 1,
  show.legend = TRUE
)
}
\arguments{
\item{df}{A `A dataframe` as generated by `coulteR::bulk_read`, with the parameter `module = "measurements"`}

\item{peaks}{A `dataframe` as generated by `coulteR::bulk_peak_detect`}

\item{seed}{Seed for random selection. Default: `123`}

\item{N}{Number of plots to be randomly selected. Default: 9}

\item{indices}{Indices of plots to be plotted. Default: `NULL`}

\item{...}{Other variables passed on to geom_line}
}
\value{
A `ggplot`-object
}
\description{
Plot target tracks
Plots tracks of selected samples. Note that file extensions and the character `#` are to be removed from sample names!
This function either plots selected samples, or selected indices, or selects and plots 9 randomly selected tracks.
}
