% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_coulter}
\alias{import_coulter}
\title{Import sample data from Coulter output.}
\usage{
import_coulter(filename, instrument = "z2", data.type = "data")
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}

\item{data.type}{One of either "data", "meta", or "full" to specify data 
to import.}

\item{instrumet}{String specifying instrument. Currently limited to Z2.}
}
\value{
A data.frame.
}
\description{
Imports sample data from exported output and converts it 
into long format. Import procedure depends on specified instrument. 
[This function is a placeholder as only Z2 import is currently
supported]. Calls import_Z2_data() and/or import_Z2_meta() as
necessary.
}
\examples{
# To do
}

