% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_z2_data}
\alias{import_z2_data}
\title{Import numeric data from text output from Coulter Z2 multisizer}
\usage{
import_z2_data(filename)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}
}
\value{
A data.frame with the following columns:
}
\description{
Imports numeric text output from a Coulter Z2 multisizer and 
converts it into long format.
}
\examples{
# To do
}

