% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_z2_meta}
\alias{import_z2_meta}
\title{Import sample description metadata from text output from 
the Coulter Z2 multisizer}
\usage{
import_z2_meta(filename)
}
\arguments{
\item{filename}{Filename or pattern for matching multiple filenames.}
}
\value{
A data.frame with columns of acquisition parameters.
}
\description{
Imports sample description metadata from text output from a Coulter Z2
multisizer and converts it into long format.
}
\examples{
# To do
}

