% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listnames_to_column}
\alias{listnames_to_column}
\title{Add ID column (list element names)}
\usage{
listnames_to_column(ls, colname = "sample")
}
\arguments{
\item{ls}{A named list of dataframes}

\item{colname}{The name of the column to add IDs to}
}
\description{
In a list of dataframes, add an ID column with the list name. Neat, if you want a dataframe instead of list object.
}
