% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-coagulated.R
\name{peak_detect}
\alias{peak_detect}
\title{Detect target peaks}
\usage{
peak_detect(df, diameter, full_df = FALSE)
}
\arguments{
\item{df}{A measurements-`module` as produced by `coulteR::read_accucomp`}

\item{diameter}{The expected diameter of particles}

\item{full_df}{Should all peaks be reported? default = `FALSE`}
}
\value{
A `dataframe` with the diameter of the peak, the diameter range, and the total number of cells in the range.
}
\description{
This function detects a specific target peak. Optionally, all peaks are reported.
}
