% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucomp-reader.R
\name{read_accucomp2}
\alias{read_accucomp2}
\title{Read Coulter Results}
\usage{
read_accucomp2(x, module = "all")
}
\arguments{
\item{x}{Path to Accucomp .XLS file}

\item{module}{Set if you want a specific module (A module). Options: all, settings, summary, size_absolute, size_summary, volume, measurements. Default: all}
}
\value{
A list of tibbles (each module is a tibble) or a single dataframe (if you specified a specific module)
}
\description{
This function reads AccuComp-tables of Coulter Counter
}
