% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucomp-reader.R
\name{read_accucomp}
\alias{read_accucomp}
\title{Read Coulter Results}
\usage{
read_accucomp(x, module = "all")
}
\arguments{
\item{x}{Path to Accucomp .XLS file}

\item{module}{Set if you want a specific module (A module). Options: one of `c("all", "settings", "summary", "size_absolute", "size_summary", "volume", "measurements")`. Default: `"all"`}
}
\value{
If `module = "all"`, a `list` of `tibbles` (each module is a tibble) or a single `tibble` (if you specified a specific module)
}
\description{
This function reads AccuComp-tables of Coulter Counter
}
